rename: term 컴포넌트 파일 이동
