+++ src/components/join/term.jsx
